cleaning/merging to master
